+++ Cargo.toml
@@ -388,9 +388,13 @@ include = ["beet-debug", "bevyhub/camera-2d", "bevyhub/ui-terminal"]
 # ]
 
 [[example]]
+name = "hello_ml_chat"
+path = "examples/ml/hello_ml_chat.rs"
+required-features = ["examples", "ml"]
+
+[[example]]
 name = "hello_ml"
 path = "examples/ml/hello_ml.rs"
-# required-features = ["examples", "ml"]
 required-features = ["ml"]
 
 [[package.metadata.scene]]
