+++ examples/ml/hello_ml_chat.rs
@@ -1,7 +1,5 @@
-//! # Hello ML
-//! A popular 'hello world' for machine learning in games is sentence similarity,
-//! where models rank the similarity of sentences.
-//! This example uses a locally run *small* language model to select the child behavior with the most similar sentence.
+//! # Hello ML Chat
+//! Like 'hello ml' but the user enters the prompt.
 use beet::examples::scenes;
 use beet::prelude::*;
 use bevy::prelude::*;
@@ -22,17 +20,26 @@ pub fn main() {
 				hello_ml,
 			),
 		)
+
 		.run();
 }
 
 
-fn hello_ml(mut commands: Commands, asset_server: Res<AssetServer>) {
-	let handle = asset_server.load("ml/default-bert.ron");
-	commands.spawn((
-		Name::new("Hello ML"),
-		RunOnAssetReady::<Bert>::new(handle),
-		SentenceBundle::with_initial("please kill the baddies"),
+fn hello_ml(
+	mut commands: Commands,
+	asset_server: Res<AssetServer>,
+
+	mut ev: EventWriter<OnLogMessage>,
+) {
+	ev.send(OnLogMessage::new(
+		"Agent: I can heal or attack, what should i do?",
 	));
+	// let handle = asset_server.load("ml/default-bert.ron");
+	// commands.spawn((
+	// 	Name::new("Hello ML"),
+	// 	RunOnAssetReady::<Bert>::new(handle),
+	// 	SentenceBundle::with_initial("please kill the baddies"),
+	// ));
 	// .with_children(|parent| {
 	// 	parent.spawn((Name::new("Heal Behavior"), Sentence::new("heal")));
 	// 	parent
